pyproject.toml
[tool.poetry]
name = "recipes-app-api"
version = "0.1.0"
description = ""
authors = ["alpden550 <alpden550@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.5"
djangorestframework = "^3.11.0"
psycopg2 = "^2.8.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
flake8-django = "^1.0.0"
autopep8 = "^1.5.2"
black = "^19.10b0"
rope = "^0.16.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
